pyproject.toml
[tool.poetry]
name = "ugc_1"
version = "0.1.0"
description = "ugc sprint 1"
authors = ["Sergey Bilibin"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv="==1.0.0"
typer = "^0.9.0"
pydantic-settings = "^2.0.2"
uvicorn = "^0.23.1"

[tool.poetry.dev-dependencies]
pytest = "==7.2.1"
isort = "==5.12.0"
flake8 = "==6.0.0"
mypy = "==1.0.0"
pre-commit = "==3.0.4"
Flake8-pyproject = "==1.2.2"
pyproject-autoflake = "==1.0.0"
clickhouse-driver = "==0.2.6"
pydantic = "==2.1.1"
pydantic-settings = "==2.0.2"
typer = "==0.9.0"
uvicorn = "==0.23.1"


[tool.autoflake]
in-place=true
expand-star-imports=true
remove-duplicate-keys=true
remove-all-unused-imports=true

[tool.black]
line-length = 120
skip-string-normalization=false

[tool.mypy]
disallow-untyped-defs=true
disallow-incomplete-defs=true
disallow-untyped-calls=true


[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=120

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 120
max-complexity = 18
select = ['B','C','E','F','W','T4']
exclude= [
    '.env.example',
    '.env',
    '.env.dev'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"